Создан контроллер облака. начинаю отладку

+++ Lesson9/CloudService/CloudService/CloudService.WebAPI/Services/Managers/ServerPoolManager.cs
@@ -1,5 +1,5 @@
 ﻿using CloudService.Interfaces;
-using CloudService.Model;
+using CloudService.Model.Models;
 using CloudService.Model.ModelsDTO;
 
 namespace CloudService.WebAPI.Services.Managers;
@@ -19,6 +19,8 @@ public class ServerPoolManager : IRepositoryAsync<ServerPoolDto>
 
     public async Task<int> AddAsync(ServerPoolDto item, CancellationToken cancel = default)
     {
+        if (item is null || item.CurrentServer is null)
+            return -1;
         ServerPool serverPool = new()
         {
             ServerId = item.CurrentServer.Id
@@ -36,37 +38,65 @@ public class ServerPoolManager : IRepositoryAsync<ServerPoolDto>
 
     public async Task<bool> DeleteAsync(ServerPoolDto item, CancellationToken cancel = default)
     {
+        if (item is null || item.CurrentServer is null)
+            return false;
         ServerPool serverPool = new()
         {
             Id = item.Id,
             ServerId = item.CurrentServer.Id
         };
+
         var result = await _serverPoolRepository.DeleteAsync(serverPool, cancel).ConfigureAwait(false);
         _logger.LogInformation(">>>Пул серверов удален");
         return result;
     }
 
-    public async Task<IEnumerable<ServerPoolDto>> GetAllAsync(CancellationToken cancel = default)
+    public async Task<IEnumerable<ServerPoolDto>?> GetAllAsync(CancellationToken cancel = default)
     {
         var serverPool = await _serverPoolRepository.GetAllAsync(cancel).ConfigureAwait(false);
+        if (serverPool is null)
+            return null;
         var serverPoolDto = new List<ServerPoolDto>();
-        foreach ( var item in serverPool)
+        foreach (var item in serverPool)
         {
             serverPoolDto.Add(new ServerPoolDto
             {
                 Id = item.Id,
-                CurrentServer
+                CurrentServer = await _serverRepository.GetByIdAsync(item.ServerId, cancel)
             });
         }
+
+        _logger.LogInformation(">>>Пул серверов получен");
+        return serverPoolDto;
     }
 
-    public Task<ServerPoolDto?> GetByIdAsync(int id, CancellationToken cancel = default)
+    public async Task<ServerPoolDto?> GetByIdAsync(int id, CancellationToken cancel = default)
     {
-        throw new NotImplementedException();
+        var serverPool = await _serverPoolRepository.GetByIdAsync(id, cancel).ConfigureAwait(false);
+        if (serverPool is null)
+            return null;
+        var serverPoolDto = new ServerPoolDto
+        {
+            Id = id,
+            CurrentServer = await _serverRepository.GetByIdAsync(serverPool.ServerId, cancel)
+        };
+
+        _logger.LogInformation(">>>Пул серверов получен");
+        return serverPoolDto;
     }
 
-    public Task<bool> UpdateAsync(ServerPoolDto item, CancellationToken cancel = default)
+    public async Task<bool> UpdateAsync(ServerPoolDto item, CancellationToken cancel = default)
     {
-        throw new NotImplementedException();
+        if (item is null || item.CurrentServer is null)
+            return false;
+        ServerPool serverPool = new()
+        {
+            Id = item.Id,
+            ServerId = item.CurrentServer.Id
+        };
+
+        var result = await _serverPoolRepository.UpdateAsync(serverPool, cancel).ConfigureAwait(false);
+        _logger.LogInformation(">>>Пул серверов изменен");
+        return result;
     }
 }
